Split PrecedenceGraph.h into separate files
