+++ src/AstToRamTranslator.cpp
@@ -35,9 +35,11 @@
 #include "ast/AstUtils.h"
 #include "ast/AstVisitor.h"
 #include "ast/TypeSystem.h"
-#include "ast/analysis/AstTypeEnvironmentAnalysis.h"
-#include "ast/analysis/AuxArityAnalysis.h"
-#include "ast/analysis/PrecedenceGraph.h"
+#include "ast/analysis/AstTypeEnvironment.h"
+#include "ast/analysis/AuxArity.h"
+#include "ast/analysis/RelationSchedule.h"
+#include "ast/analysis/SCCGraph.h"
+#include "ast/analysis/TopologicallySortedSCCGraph.h"
 #include "ram/RamCondition.h"
 #include "ram/RamExpression.h"
 #include "ram/RamNode.h"
