+++ src/ast/analysis/ComponentLookup.h
@@ -8,7 +8,7 @@
 
 /************************************************************************
  *
- * @file ComponentLookupAnalysis.h
+ * @file ComponentLookup.h
  *
  ***********************************************************************/
