+++ src/ast/transform/AstSemanticChecker.cpp
@@ -37,11 +37,14 @@
 #include "ast/AstUtils.h"
 #include "ast/AstVisitor.h"
 #include "ast/TypeSystem.h"
-#include "ast/analysis/AstGroundAnalysis.h"
-#include "ast/analysis/AstIOTypeAnalysis.h"
-#include "ast/analysis/AstTypeAnalysis.h"
-#include "ast/analysis/AstTypeEnvironmentAnalysis.h"
+#include "ast/analysis/AstGround.h"
+#include "ast/analysis/AstIOType.h"
+#include "ast/analysis/AstType.h"
+#include "ast/analysis/AstTypeEnvironment.h"
 #include "ast/analysis/PrecedenceGraph.h"
+#include "ast/analysis/RecursiveClauses.h"
+#include "ast/analysis/RelationSchedule.h"
+#include "ast/analysis/SCCGraph.h"
 #include "souffle/BinaryConstraintOps.h"
 #include "souffle/RamTypes.h"
 #include "souffle/utility/ContainerUtil.h"
