+++ src/ast/transform/ComponentInstantiation.cpp
@@ -28,7 +28,7 @@
 #include "ast/AstTranslationUnit.h"
 #include "ast/AstType.h"
 #include "ast/AstVisitor.h"
-#include "ast/analysis/ComponentLookupAnalysis.h"
+#include "ast/analysis/ComponentLookup.h"
 #include "souffle/utility/StringUtil.h"
 #include <algorithm>
 #include <cstddef>
