+++ src/ast/transform/MaterializeAggregationQueries.cpp
@@ -27,9 +27,9 @@
 #include "ast/AstUtils.h"
 #include "ast/AstVisitor.h"
 #include "ast/TypeSystem.h"
-#include "ast/analysis/AstGroundAnalysis.h"
-#include "ast/analysis/AstTypeAnalysis.h"
-#include "ast/analysis/AstTypeEnvironmentAnalysis.h"
+#include "ast/analysis/AstGround.h"
+#include "ast/analysis/AstType.h"
+#include "ast/analysis/AstTypeEnvironment.h"
 #include "souffle/RamTypes.h"
 #include "souffle/utility/MiscUtil.h"
 #include "souffle/utility/StringUtil.h"
