+++ src/ast/transform/PartitionBodyLiterals.cpp
@@ -24,7 +24,6 @@
 #include "ast/AstTranslationUnit.h"
 #include "ast/AstVisitor.h"
 #include "souffle/utility/MiscUtil.h"
-#include <algorithm>
 #include <functional>
 #include <memory>
 #include <ostream>
