+++ src/ast/transform/ReduceExistentials.cpp
@@ -26,7 +26,7 @@
 #include "ast/AstTranslationUnit.h"
 #include "ast/AstUtils.h"
 #include "ast/AstVisitor.h"
-#include "ast/analysis/AstIOTypeAnalysis.h"
+#include "ast/analysis/AstIOType.h"
 #include "souffle/utility/MiscUtil.h"
 #include <functional>
 #include <memory>
