+++ src/ast/transform/RemoveEmptyRelations.cpp
@@ -22,7 +22,7 @@
 #include "ast/AstTranslationUnit.h"
 #include "ast/AstUtils.h"
 #include "ast/AstVisitor.h"
-#include "ast/analysis/AstIOTypeAnalysis.h"
+#include "ast/analysis/AstIOType.h"
 #include "souffle/utility/MiscUtil.h"
 #include <algorithm>
 #include <memory>
