+++ src/tests/ast_utils_test.cpp
@@ -28,7 +28,7 @@
 #include "ast/AstQualifiedName.h"
 #include "ast/AstTranslationUnit.h"
 #include "ast/AstUtils.h"
-#include "ast/analysis/AstGroundAnalysis.h"
+#include "ast/analysis/AstGround.h"
 #include "souffle/BinaryConstraintOps.h"
 #include "souffle/utility/StringUtil.h"
 #include <algorithm>
